refactor(metadata): clarify editor state ownership
